requirements.txt
pandas
numpy
pydantic
pydantic-settings
openai
httpx
scikit-learn
python-dotenv
tqdm
